requirements.txt
asgiref==3.5.1
colorama==0.4.4
dj-database-url==0.5.0
Django==4.0.4
django-appconf==1.0.5
django-phonenumber-field==6.1.0
django-select2==7.10.0
gunicorn==20.1.0
numpy==1.22.4
opencv-python==4.5.5.64
phonenumberslite==8.12.48
Pillow==9.1.0
psycopg2==2.9.3
python-barcode==0.13.1
pyzbar==0.1.9
qrcode==7.3.1
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.2.0
